requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-ckeditor==6.7.0
redis==5.0.1
django-redis==5.4.0
googletrans==3.1.0a0
python-dotenv==1.0.1
pytest==8.0.0
pytest-django==4.8.0
flake8==7.0.0
gunicorn==21.2.0